modified based on cr
