+++ fluss-flink/fluss-flink-common/src/test/java/com/alibaba/fluss/flink/security/acl/FlinkAuthorizationITCase.java
@@ -27,7 +27,10 @@ import com.alibaba.fluss.metadata.TablePath;
 import com.alibaba.fluss.security.acl.FlussPrincipal;
 import com.alibaba.fluss.security.acl.OperationType;
 import com.alibaba.fluss.security.acl.Resource;
+import com.alibaba.fluss.security.auth.sasl.jaas.LoginManager;
 import com.alibaba.fluss.server.testutils.FlussClusterExtension;
+import com.alibaba.fluss.utils.ParentResourceBlockingClassLoader;
+import com.alibaba.fluss.utils.TemporaryClassLoaderContext;
 
 import org.apache.commons.lang3.RandomUtils;
 import org.apache.flink.table.api.EnvironmentSettings;
@@ -43,6 +46,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import java.net.URL;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collections;
@@ -383,6 +387,28 @@ abstract class FlinkAuthorizationITCase extends AbstractTestBase {
                 tBatchEnv,
                 limitScanSql,
                 Arrays.asList("+I[1, beijing, zhangsan]", "+I[2, shanghai, lisi]"));
+    }
+
+    // this test is to mock `--jar` which is not loaded by the app classloader.
+    @Test
+    void testNotIncludedInThreadContextClassloader() throws Exception {
+        try (TemporaryClassLoaderContext ignored =
+                TemporaryClassLoaderContext.of(new ParentResourceBlockingClassLoader(new URL[0]))) {
+            // clear the cache of login context to make sure load the class again.
+            LoginManager.closeAll();
+            tEnv.executeSql(
+                            String.format(
+                                    "create catalog test_classloader_catalog with ('type' = 'fluss', "
+                                            + "'bootstrap.servers' = '%s',"
+                                            + "'client.security.protocol' = 'sasl',"
+                                            + "'client.security.sasl.mechanism' = 'PLAIN', \n"
+                                            + "'client.security.sasl.username' = 'guest', \n"
+                                            + "'client.security.sasl.password' = 'password2' \n"
+                                            + ")",
+                                    String.join(
+                                            ",", clientConf.get(ConfigOptions.BOOTSTRAP_SERVERS))))
+                    .await();
+        }
     }
 
     void addAcl(Resource resource, OperationType operationType)
